requirements.txt
Flask>=3.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
pandas>=2.1.0
numpy>=1.26.0
alpaca-py>=0.21.0
plotly>=5.17.0
pytest
